Refactor (file) descriptors
