+++ src/libc/putchar.c
@@ -1,11 +1,11 @@
 #include <stdio.h>
 
 #ifdef __is_libc
-#include <fs/fd.h>
+#include <proc/descriptor.h>
 #include <sys/syscall.h>
 
 void putchar(char c)
 {
-  write(FD_STDOUT, &c, 1);
+  write(STDOUT, &c, 1);
 }
 #endif
